Rename domains dashboard navigation action for consistency
