+++ WordPress/src/main/java/org/wordpress/android/ui/domains/DomainsDashboardNavigationAction.kt
@@ -1,0 +1,8 @@
+package org.wordpress.android.ui.domains
+
+import org.wordpress.android.fluxc.model.SiteModel
+
+sealed class DomainsDashboardNavigationAction {
+    data class GetDomain(val site: SiteModel) : DomainsDashboardNavigationAction()
+    data class OpenManageDomains(val url: String) : DomainsDashboardNavigationAction()
+}
